Handling case for filtering by ID with GET
For some reason, when you include `{id: objectID}` as a filter parameter, Sails will return a single object, rather than an array with one item in it.  For any other queries that return only one item, Sails returns an array.

Seems that angular doesn't return the result if it's an object, but it does if it's an array.

This change just checks to see if a single object is returned and if so, it pushes it into a new array.